+++ cg/cli/set/cases.py
