+++ tests/cli/set/test_cli_set_case.py
@@ -1,4 +1,4 @@
-"""This script tests the cli methods to set values of cases in status-db."""
+"""This script tests the cli methods to set values of a case in status-db."""
 from click.testing import CliRunner
 
 from cg.cli.set.family import case
